python/pyproject.toml
[tool.poetry]
name = "python"
version = "0.1.0"
description = ""
authors = ["Jojo <jonathan.silva@birdie.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic-factories = "^1.17.0"
matplotlib = "^3.6.3"
loguru = "^0.6.0"
google-cloud-logging = "^3.5.0"
structlog = "^22.3.0"
python-json-logger = "^2.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
